refactor(mobile): extract queue provider session realtime effect
